refactor plot tests
